requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
cryptography==41.0.2
dataclass-wizard==0.22.2
dj-database-url==2.0.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
ed25519==1.5
gunicorn==21.2.0
inflection==0.5.1
jedi==0.18.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jwt==1.3.1
packaging==23.1
parso==0.8.3
Pillow==10.0.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
Token==1.8.6
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
